fix nav menu
