+++ src/components/Navigation/Navigation.tsx
@@ -1,26 +1,19 @@
 /* eslint-disable react/no-children-prop */
-import { NavLinkRenderProps } from 'react-router-dom';
 import { useEffect, useState } from 'react';
-import { shallowEqual } from 'react-redux';
 import { useTranslation } from 'react-i18next';
 
 import Menu from '../../assets/Menu.svg?react';
-import Auth from '../../assets/Auth.svg?react';
 import Close from '../../assets/Close.svg?react';
-import Cart from '../../assets/Shopping.svg?react';
-import Favorites from '../../assets/Favourites.svg?react';
 import Logo from '../Logo/Logo';
 import NavigationLink from '../../UI/NavLink/NavigationLink';
 import { ThemeSwitcher } from '../ThemeSwitcher/ThemeSwitcher';
 import { LanguageSwitcher } from '../LanguageSwitcher/LanguageSwitcher';
 
-import { IconLink } from './IconLink/IconLink';
 import s from './Navigation.module.scss';
+import { IconLinks } from './IconLinks/IconLinks';
 
 import { IconButton } from '@/UI/IconButton/IconButton';
 import { PATH } from '@/constants/path';
-import { useAppSelector } from '@/redux/hooks';
-import { selectCart, selectFavorites, selectUser } from '@/redux/selectors';
 import { allowScroll, blockScroll } from '@/utils/scroll';
 
 const navLinks = {
@@ -30,13 +23,9 @@ const navLinks = {
   Accessories: PATH.ACCESSORIES,
 };
 
-const navLinkStyle = ({ isActive }: NavLinkRenderProps) => `${s.link} ${isActive ? s.active_link : ''}`;
-
 const Navigation = () => {
   const [isOpenMobileMenu, setIsOpenMobileMenu] = useState(false);
-  const user = useAppSelector(selectUser, shallowEqual);
-  const cartList = useAppSelector(selectCart, shallowEqual);
-  const favoritesList = useAppSelector(selectFavorites, shallowEqual);
+  const closeBurgerMenu = () => setIsOpenMobileMenu(false);
   const { t } = useTranslation('navigation');
 
   useEffect(() => {
@@ -60,37 +49,15 @@ const Navigation = () => {
               key={label}
               to={path}
               label={t(label)}
-              className={navLinkStyle}
-              handleClick={() => setIsOpenMobileMenu(false)}
+              className={({ isActive }) => `${s.link} ${isActive ? s.active_link : ''}`}
+              handleClick={closeBurgerMenu}
             />
           ))}
         </div>
-        <div className={s.navigation__icons}>
-          {!user && (
-            <IconLink
-              icon={<Auth />}
-              path={`/auth${PATH.SIGN_IN}`}
-              counter={0}
-              handleClick={() => setIsOpenMobileMenu(false)}
-            />
-          )}
-
-          <IconLink
-            icon={<Cart />}
-            path={PATH.CART}
-            counter={cartList.length}
-            handleClick={() => setIsOpenMobileMenu(false)}
-          />
-          <IconLink
-            icon={<Favorites />}
-            path={PATH.FAVORITES}
-            counter={favoritesList.length}
-            handleClick={() => setIsOpenMobileMenu(false)}
-          />
-        </div>
+        <IconLinks handleClick={closeBurgerMenu} />
       </nav>
 
-      <div className={s.wrapper}>
+      <div className={s.switches}>
         <ThemeSwitcher />
         <LanguageSwitcher />
